feat: 반환되는 DTO에서 불필요한 Attributes 제거

+++ src/main/java/com/ns/solve/domain/dto/board/BoardSummary.java
@@ -12,5 +12,5 @@ public class BoardSummary {
     private String title;
     private String creator;
     private LocalDateTime lastModified;
-    private Integer commentCount;
+    private Long commentCount;
 }
